split app.tsx into smaller components
