+++ src/constants/categories.ts
@@ -1,0 +1,10 @@
+import { Category } from '../types/models';
+
+export const CATEGORIES: Category[] = [
+  'Hair Care',
+  'Body Care',
+  'Makeup',
+  'Fragrance',
+  'Skin Care',
+  'Other',
+];
